+++ Services/Live/V1/Model/ShowDomainStreamBackupRequest.cs
@@ -1,0 +1,72 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.Live.V1.Model
+{
+    /// <summary>
+    /// Request Object
+    /// </summary>
+    public class ShowDomainStreamBackupRequest 
+    {
+
+        /// <summary>
+        /// **参数解释**： 直播推流域名 **约束限制**： 不涉及 **取值范围**： 字符长度为[1-255]位 **默认取值**： 不涉及 
+        /// </summary>
+        [SDKProperty("publish_domain", IsQuery = true)]
+        [JsonProperty("publish_domain", NullValueHandling = NullValueHandling.Ignore)]
+        public string PublishDomain { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ShowDomainStreamBackupRequest {\n");
+            sb.Append("  publishDomain: ").Append(PublishDomain).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ShowDomainStreamBackupRequest);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ShowDomainStreamBackupRequest input)
+        {
+            if (input == null) return false;
+            if (this.PublishDomain != input.PublishDomain || (this.PublishDomain != null && !this.PublishDomain.Equals(input.PublishDomain))) return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                if (this.PublishDomain != null) hashCode = hashCode * 59 + this.PublishDomain.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}